lint: reorganize package
The lint_test.go file was getting rather large. Split out the custom
"checkers" into their own files.

Release note: None
