+++ pkg/testutils/lint/lint_test.go
@@ -20,12 +20,8 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"go/ast"
 	"go/build"
 	"go/parser"
-	"go/token"
-	"go/types"
-	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -1304,381 +1300,4 @@ func TestLint(t *testing.T) {
 		}
 	})
 
-}
-
-type miscChecker struct{}
-
-func (*miscChecker) Name() string {
-	return "misccheck"
-}
-
-func (*miscChecker) Prefix() string {
-	return "CR"
-}
-
-func (*miscChecker) Init(*lint.Program) {}
-
-func (*miscChecker) Funcs() map[string]lint.Func {
-	return map[string]lint.Func{
-		"FloatToUnsigned": checkConvertFloatToUnsigned,
-		"Unconvert":       checkUnconvert,
-	}
-}
-
-func forAllFiles(j *lint.Job, fn func(node ast.Node) bool) {
-	for _, f := range j.Program.Files {
-		if !lint.IsGenerated(f) {
-			ast.Inspect(f, fn)
-		}
-	}
-}
-
-// @ianlancetaylor via golang-nuts[0]:
-//
-// For the record, the spec says, in https://golang.org/ref/spec#Conversions:
-// "In all non-constant conversions involving floating-point or complex
-// values, if the result type cannot represent the value the conversion
-// succeeds but the result value is implementation-dependent."  That is the
-// case that applies here: you are converting a negative floating point number
-// to uint64, which can not represent a negative value, so the result is
-// implementation-dependent.  The conversion to int64 works, of course. And
-// the conversion to int64 and then to uint64 succeeds in converting to int64,
-// and when converting to uint64 follows a different rule: "When converting
-// between integer types, if the value is a signed integer, it is sign
-// extended to implicit infinite precision; otherwise it is zero extended. It
-// is then truncated to fit in the result type's size."
-//
-// So, basically, don't convert a negative floating point number to an
-// unsigned integer type.
-//
-// [0] https://groups.google.com/d/msg/golang-nuts/LH2AO1GAIZE/PyygYRwLAwAJ
-//
-// TODO(tamird): upstream this.
-func checkConvertFloatToUnsigned(j *lint.Job) {
-	forAllFiles(j, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			return true
-		}
-		castType, ok := j.Program.Info.TypeOf(call.Fun).(*types.Basic)
-		if !ok {
-			return true
-		}
-		if castType.Info()&types.IsUnsigned == 0 {
-			return true
-		}
-		for _, arg := range call.Args {
-			argType, ok := j.Program.Info.TypeOf(arg).(*types.Basic)
-			if !ok {
-				continue
-			}
-			if argType.Info()&types.IsFloat == 0 {
-				continue
-			}
-			j.Errorf(arg, "do not convert a floating point number to an unsigned integer type")
-		}
-		return true
-	})
-}
-
-func walkForStmts(n ast.Node, fn func(ast.Stmt) bool) bool {
-	fr, ok := n.(*ast.ForStmt)
-	if !ok {
-		return true
-	}
-	return walkStmts(fr.Body.List, fn)
-}
-
-func walkSelectStmts(n ast.Node, fn func(ast.Stmt) bool) bool {
-	sel, ok := n.(*ast.SelectStmt)
-	if !ok {
-		return true
-	}
-	return walkStmts(sel.Body.List, fn)
-}
-
-func walkStmts(stmts []ast.Stmt, fn func(ast.Stmt) bool) bool {
-	for _, stmt := range stmts {
-		if !fn(stmt) {
-			return false
-		}
-	}
-	return true
-}
-
-// timerChecker assures that timeutil.Timer objects are used correctly, to
-// avoid race conditions and deadlocks. These timers require callers to set
-// their Read field to true when their channel has been received on. If this
-// field is not set and the timer's Reset method is called, we will deadlock.
-// This lint assures that the Read field is set in the most common case where
-// Reset is used, within a for-loop where each iteration blocks on a select
-// statement. The timers are usually used as timeouts on these select
-// statements, and need to be reset after each iteration.
-//
-// for {
-//   timer.Reset(...)
-//   select {
-//     case <-timer.C:
-//       timer.Read = true   <--  lint verifies that this line is present
-//     case ...:
-//   }
-// }
-//
-type timerChecker struct {
-	timerType types.Type
-}
-
-func (*timerChecker) Name() string {
-	return "timercheck"
-}
-
-func (*timerChecker) Prefix() string {
-	return "T"
-}
-
-const timerChanName = "C"
-
-func (m *timerChecker) Init(program *lint.Program) {
-	timeutilPkg := program.Prog.Package("github.com/cockroachdb/cockroach/pkg/util/timeutil")
-	if timeutilPkg == nil {
-		log.Fatal("timeutil package not found")
-	}
-	timerObject := timeutilPkg.Pkg.Scope().Lookup("Timer")
-	if timerObject == nil {
-		log.Fatal("timeutil.Timer type not found")
-	}
-	m.timerType = timerObject.Type()
-
-	func() {
-		if typ, ok := m.timerType.Underlying().(*types.Struct); ok {
-			for i := 0; i < typ.NumFields(); i++ {
-				if typ.Field(i).Name() == timerChanName {
-					return
-				}
-			}
-		}
-		log.Fatalf("no field called %q in type %s", timerChanName, m.timerType)
-	}()
-}
-
-func (m *timerChecker) Funcs() map[string]lint.Func {
-	return map[string]lint.Func{
-		"TimeutilTimerRead": m.checkSetTimeutilTimerRead,
-	}
-}
-
-func (m *timerChecker) selectorIsTimer(s *ast.SelectorExpr, info *types.Info) bool {
-	selTyp := info.TypeOf(s.X)
-	if ptr, ok := selTyp.(*types.Pointer); ok {
-		selTyp = ptr.Elem()
-	}
-	return selTyp == m.timerType
-}
-
-func (m *timerChecker) checkSetTimeutilTimerRead(j *lint.Job) {
-	forAllFiles(j, func(n ast.Node) bool {
-		return walkForStmts(n, func(s ast.Stmt) bool {
-			return walkSelectStmts(s, func(s ast.Stmt) bool {
-				comm, ok := s.(*ast.CommClause)
-				if !ok || comm.Comm == nil /* default: */ {
-					return true
-				}
-
-				// if receiving on a timer's C chan.
-				var unary ast.Expr
-				switch v := comm.Comm.(type) {
-				case *ast.AssignStmt:
-					// case `now := <-timer.C:`
-					unary = v.Rhs[0]
-				case *ast.ExprStmt:
-					// case `<-timer.C:`
-					unary = v.X
-				default:
-					return true
-				}
-				chanRead, ok := unary.(*ast.UnaryExpr)
-				if !ok || chanRead.Op != token.ARROW {
-					return true
-				}
-				selector, ok := chanRead.X.(*ast.SelectorExpr)
-				if !ok {
-					return true
-				}
-				if !m.selectorIsTimer(selector, j.Program.Info) {
-					return true
-				}
-				selectorName := fmt.Sprint(selector.X)
-				if selector.Sel.String() != timerChanName {
-					return true
-				}
-
-				// Verify that the case body contains `timer.Read = true`.
-				noRead := walkStmts(comm.Body, func(s ast.Stmt) bool {
-					assign, ok := s.(*ast.AssignStmt)
-					if !ok || assign.Tok != token.ASSIGN {
-						return true
-					}
-					for i := range assign.Lhs {
-						l, r := assign.Lhs[i], assign.Rhs[i]
-
-						// if assignment to correct field in timer.
-						assignSelector, ok := l.(*ast.SelectorExpr)
-						if !ok {
-							return true
-						}
-						if !m.selectorIsTimer(assignSelector, j.Program.Info) {
-							return true
-						}
-						if fmt.Sprint(assignSelector.X) != selectorName {
-							return true
-						}
-						if assignSelector.Sel.String() != "Read" {
-							return true
-						}
-
-						// if assigning `true`.
-						val, ok := r.(*ast.Ident)
-						if !ok {
-							return true
-						}
-						if val.String() == "true" {
-							// returning false will short-circuit walkStmts and assign
-							// noRead to false instead of the default value of true.
-							return false
-						}
-					}
-					return true
-				})
-				if noRead {
-					j.Errorf(comm, "must set timer.Read = true after reading from timer.C (see timeutil/timer.go)")
-				}
-				return true
-			})
-		})
-	})
-}
-
-// Adapted from https://github.com/mdempsky/unconvert/blob/beb68d938016d2dec1d1b078054f4d3db25f97be/unconvert.go#L371-L414.
-func checkUnconvert(j *lint.Job) {
-	forAllFiles(j, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			return true
-		}
-		if len(call.Args) != 1 || call.Ellipsis != token.NoPos {
-			return true
-		}
-		ft, ok := j.Program.Info.Types[call.Fun]
-		if !ok {
-			j.Errorf(call.Fun, "missing type")
-			return true
-		}
-		if !ft.IsType() {
-			// Function call; not a conversion.
-			return true
-		}
-		at, ok := j.Program.Info.Types[call.Args[0]]
-		if !ok {
-			j.Errorf(call.Args[0], "missing type")
-			return true
-		}
-		if !types.Identical(ft.Type, at.Type) {
-			// A real conversion.
-			return true
-		}
-		if isUntypedValue(call.Args[0], j.Program.Info) {
-			// Workaround golang.org/issue/13061.
-			return true
-		}
-		// Adapted from https://github.com/mdempsky/unconvert/blob/beb68d938016d2dec1d1b078054f4d3db25f97be/unconvert.go#L416-L430.
-		//
-		// cmd/cgo generates explicit type conversions that
-		// are often redundant when introducing
-		// _cgoCheckPointer calls (issue #16).  Users can't do
-		// anything about these, so skip over them.
-		if ident, ok := call.Fun.(*ast.Ident); ok {
-			if ident.Name == "_cgoCheckPointer" {
-				return false
-			}
-		}
-		j.Errorf(n, "unnecessary conversion")
-		return true
-	})
-}
-
-// Cribbed from https://github.com/mdempsky/unconvert/blob/beb68d938016d2dec1d1b078054f4d3db25f97be/unconvert.go#L557-L607.
-func isUntypedValue(n ast.Expr, info *types.Info) bool {
-	switch n := n.(type) {
-	case *ast.BinaryExpr:
-		switch n.Op {
-		case token.SHL, token.SHR:
-			// Shifts yield an untyped value if their LHS is untyped.
-			return isUntypedValue(n.X, info)
-		case token.EQL, token.NEQ, token.LSS, token.GTR, token.LEQ, token.GEQ:
-			// Comparisons yield an untyped boolean value.
-			return true
-		case token.ADD, token.SUB, token.MUL, token.QUO, token.REM,
-			token.AND, token.OR, token.XOR, token.AND_NOT,
-			token.LAND, token.LOR:
-			return isUntypedValue(n.X, info) && isUntypedValue(n.Y, info)
-		}
-	case *ast.UnaryExpr:
-		switch n.Op {
-		case token.ADD, token.SUB, token.NOT, token.XOR:
-			return isUntypedValue(n.X, info)
-		}
-	case *ast.BasicLit:
-		// Basic literals are always untyped.
-		return true
-	case *ast.ParenExpr:
-		return isUntypedValue(n.X, info)
-	case *ast.SelectorExpr:
-		return isUntypedValue(n.Sel, info)
-	case *ast.Ident:
-		if obj, ok := info.Uses[n]; ok {
-			if obj.Pkg() == nil && obj.Name() == "nil" {
-				// The universal untyped zero value.
-				return true
-			}
-			if b, ok := obj.Type().(*types.Basic); ok && b.Info()&types.IsUntyped != 0 {
-				// Reference to an untyped constant.
-				return true
-			}
-		}
-	case *ast.CallExpr:
-		if b, ok := asBuiltin(n.Fun, info); ok {
-			switch b.Name() {
-			case "real", "imag":
-				return isUntypedValue(n.Args[0], info)
-			case "complex":
-				return isUntypedValue(n.Args[0], info) && isUntypedValue(n.Args[1], info)
-			}
-		}
-	}
-
-	return false
-}
-
-// Cribbed from https://github.com/mdempsky/unconvert/blob/beb68d938016d2dec1d1b078054f4d3db25f97be/unconvert.go#L609-L630.
-func asBuiltin(n ast.Expr, info *types.Info) (*types.Builtin, bool) {
-	for {
-		paren, ok := n.(*ast.ParenExpr)
-		if !ok {
-			break
-		}
-		n = paren.X
-	}
-
-	ident, ok := n.(*ast.Ident)
-	if !ok {
-		return nil, false
-	}
-
-	obj, ok := info.Uses[ident]
-	if !ok {
-		return nil, false
-	}
-
-	b, ok := obj.(*types.Builtin)
-	return b, ok
 }
